Adds support for specifying color 0 (Not Applicable) as the color for the part, indicating that any color for the part is acceptable. This matches the behavior when specifying color 0 in a BrickLink Wanted List.  This is very useful for getting the best price for builds containing internal or structural elements where the color is not important.
When color 0 is specifed, Brickficiency will first find all available
colors for the part and then fetch the price guide information for each
color in turn. Each store's lowest price and the corresponding color will
be stored against color 0 in that store's databse entry.

When finding store combinations, the cheapest color for that particular
store will be used.